pyproject.toml
[tool.poetry]
name = "cover-letter-generator"
version = "0.1.0"
description = ""
authors = ["rcparent17 <reilly.c.parent@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pdfkit = "^1.0.0"
pypdf = "^6.0.0"
pytest = "^8.4.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
